move namespaces around for BuildSummary related types
