+++ src/ProjectSystemTools/BuildLogging/Model/BuildSummary.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.IO;
 
-namespace Microsoft.VisualStudio.ProjectSystem.Tools.Providers
+namespace Microsoft.VisualStudio.ProjectSystem.Tools.BuildLogging.Model
 {
     /// <summary>
     /// Immutable Type
